backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pymongo[snappy,gssapi]==4.13.0
motor==3.7.1
python-dotenv==1.0.0
transformers==4.30.2
torch==2.0.1
langchain==0.0.208
dnspython==2.7.0
huggingface_hub
starlette==0.27.0
pdfplumber==0.9.0 
python-multipart==0.0.6
